Improving summon logic
Note: it seems like there's a bug, but there's not.

If you right click the "register block" multiple times while horses are
registered, it will add a bunch of "copies" of the registered horse.
But you can only spawn 1 real horse from them before they stop working.

This is not actually a bug because the mod won't be used this way. We
will spawn stables in the world which will have fake horses placed at
specific positions, and each one of those horses will have a unique
integer "index" ID stored on them. Which means one horse will be
strongly associated to (for example) the "fifth" registered horse, and
none of the other horses at the stable will represent that horse.

In other words: We will not be spawning fake horses programatically.
So we don't need to worry about duplicate fakes.